feat: fix api and user crud

+++ server/services/product.service.ts
@@ -1,20 +1,22 @@
 import prisma from './prisma'
-import type { Product } from '~/generated/prisma'
+import type { IService } from './IService'
+import type { CreateProductDto, ProductDto, UpdateProductDto } from '~/types/dto/product.dto'
 
-export class ProductService {
+
+export class ProductService implements IService<ProductDto> {
   // 建立產品
-  async create(data: {
-    name: string
-    description?: string
-  }) {
+  async create(data: CreateProductDto) {
     return await prisma.product.create({
       data
     })
   }
 
   // 取得所有產品
-  async findAll() {
-    return await prisma.product.findMany({
+  async findAll(page: number, pageSize: number) {
+    const total = await prisma.product.count()
+    const products = await prisma.product.findMany({
+      skip: (page - 1) * pageSize,
+      take: pageSize,
       include: {
         units: {
           include: {
@@ -23,6 +25,10 @@ export class ProductService {
         }
       }
     })
+    return {
+      data: products,
+      pageInfo: { page, pageSize, total }
+    }
   }
 
   // 取得單一產品
@@ -41,10 +47,16 @@ export class ProductService {
   }
 
   // 更新產品
-  async update(id: number, data: {
-    name?: string
-    description?: string
-  }) {
+  async update(id: number, data: UpdateProductDto) {
+    const product = await prisma.product.findUnique({
+      where: { id }
+    })
+    if (!product) {
+      throw createError({
+        statusCode: 404,
+        statusMessage: 'Product not found'
+      })
+    }
     return await prisma.product.update({
       where: { id },
       data
@@ -53,7 +65,16 @@ export class ProductService {
 
   // 刪除產品
   async delete(id: number) {
-    return await prisma.product.delete({
+    const product = await prisma.product.findUnique({
+      where: { id }
+    })
+    if (!product) {
+      throw createError({
+        statusCode: 404,
+        statusMessage: 'Product not found'
+      })
+    }
+    await prisma.product.delete({
       where: { id }
     })
   }
